+++ src/main/model/Rivian.java
@@ -1,7 +1,9 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Rivian extends Stock {
-    public Rivian(String company, int price) {
+    public Rivian(String company, BigDecimal price) {
         super(company, price);
     }
 }
