+++ src/main/model/Transaction.java
@@ -1,10 +1,12 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Transaction {
     private Stock stock;
-    private int price;
+    private BigDecimal price;
 
-    public Transaction(Stock stock, int price) {
+    public Transaction(Stock stock, BigDecimal price) {
         this.stock = stock;
         this.price = price;
     }
